requirements.txt
# Snowflake connector and core dependencies
snowflake-connector-python>=3.7.0
python-dotenv>=1.0.0
click>=8.0.0

# Additional utilities
requests>=2.31.0
pydantic>=2.0.0
rich>=13.0.0